Update Visual C++ property sheets
The last commit (Add a preconfigured gio/gnetworking.h for Windows) has to
be split into two as git am does not like a patch that deals with files
that have different line feeds.

This updates the property sheets to use the pre-configured
gio/gnetworking.h during the build process.

https://bugzilla.gnome.org/show_bug.cgi?id=690163